export `ConditionalSimplify` and `ConditionalSimplifyDeep`
